refactoring to reuse KQL filter from observability
