+++ x-pack/plugins/apm/public/components/alerting/rule_types/transaction_error_rate_rule_type/index.tsx
@@ -47,7 +47,7 @@ import {
   LoadingState,
   NoDataState,
 } from '../../ui_components/chart_preview/chart_preview_helper';
-import { ApmRuleKqlFilter } from '../../ui_components/apm_rule_kql_filter/rule_kql_filter';
+import { ApmRuleKqlFilter } from '../../ui_components/apm_rule_kql_filter';
 
 export interface ErrorRateRuleParams {
   windowSize?: number;
