--- x-pack/plugins/apm/public/components/alerting/ui_components/apm_rule_kql_filter/with_kuery_autocompletion.tsx
@@ -1,123 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { DataViewBase } from '@kbn/es-query';
-import {
-  withKibana,
-  KibanaReactContextValue,
-  KibanaServices,
-} from '@kbn/kibana-react-plugin/public';
-import type { DataView } from '@kbn/data-views-plugin/public';
-import { QuerySuggestion } from '@kbn/unified-search-plugin/public';
-import { ApmPluginStartDeps } from '../../../../plugin';
-import { RendererFunction } from './typed_react';
-
-interface WithKueryAutocompletionLifecycleProps {
-  kibana: KibanaReactContextValue<ApmPluginStartDeps & KibanaServices>;
-  children: RendererFunction<{
-    isLoadingSuggestions: boolean;
-    loadSuggestions: (
-      expression: string,
-      cursorPosition: number,
-      maxSuggestions?: number
-    ) => void;
-    suggestions: QuerySuggestion[];
-  }>;
-  indexPattern?: DataViewBase;
-}
-
-interface WithKueryAutocompletionLifecycleState {
-  // lacking cancellation support in the autocompletion api,
-  // this is used to keep older, slower requests from clobbering newer ones
-  currentRequest: {
-    expression: string;
-    cursorPosition: number;
-  } | null;
-  suggestions: QuerySuggestion[];
-}
-
-class WithKueryAutocompletionComponent extends React.Component<
-  WithKueryAutocompletionLifecycleProps,
-  WithKueryAutocompletionLifecycleState
-> {
-  public readonly state: WithKueryAutocompletionLifecycleState = {
-    currentRequest: null,
-    suggestions: [],
-  };
-
-  public render() {
-    const { currentRequest, suggestions } = this.state;
-
-    return this.props.children({
-      isLoadingSuggestions: currentRequest !== null,
-      loadSuggestions: this.loadSuggestions,
-      suggestions,
-    });
-  }
-
-  private loadSuggestions = async (
-    expression: string,
-    cursorPosition: number,
-    maxSuggestions?: number,
-    transformSuggestions?: (s: QuerySuggestion[]) => QuerySuggestion[]
-  ) => {
-    const { indexPattern } = this.props;
-    const language = 'kuery';
-    const hasQuerySuggestions =
-      this.props.kibana.services.unifiedSearch.autocomplete.hasQuerySuggestions(
-        language
-      );
-
-    if (!hasQuerySuggestions) {
-      return;
-    }
-
-    this.setState({
-      currentRequest: {
-        expression,
-        cursorPosition,
-      },
-      suggestions: [],
-    });
-
-    const suggestions =
-      (await this.props.kibana.services.unifiedSearch.autocomplete.getQuerySuggestions(
-        {
-          language,
-          query: expression,
-          selectionStart: cursorPosition,
-          selectionEnd: cursorPosition,
-          indexPatterns: [indexPattern as DataView],
-          boolFilter: [],
-        }
-      )) || [];
-
-    const transformedSuggestions = transformSuggestions
-      ? transformSuggestions(suggestions)
-      : suggestions;
-
-    this.setState((state) =>
-      state.currentRequest &&
-      state.currentRequest.expression !== expression &&
-      state.currentRequest.cursorPosition !== cursorPosition
-        ? state // ignore this result, since a newer request is in flight
-        : {
-            ...state,
-            currentRequest: null,
-            suggestions: maxSuggestions
-              ? transformedSuggestions.slice(0, maxSuggestions)
-              : transformedSuggestions,
-          }
-    );
-  };
-}
-
-export const WithKueryAutocompletion =
-  withKibana<WithKueryAutocompletionLifecycleProps>(
-    WithKueryAutocompletionComponent
-  );
