+++ x-pack/plugins/observability/public/components/rule_kql_filter/autocomplete_field/index.ts
